myfly/requirements.txt
cx_Oracle==7.2.1
sqlalchemy==1.3.6
MySQL-connector-python=8.0.17